Move node commands to specific files.
